魔能发送器同步，魔能残卷制作中

+++ Content/Items/Magike/Filters/DiffractionFilters/MagicCrystalDiffractionFilter.cs
@@ -38,6 +38,6 @@ namespace Coralite.Content.Items.Magike.Filters.DiffractionFilters
 
         public override int ItemType => ModContent.ItemType<MagicCrystalDiffractionFilter>();
 
-        public override int ConnectMaxBonus => 1;
+        public override byte ConnectMaxBonus => 1;
     }
 }
